+++ packages/build-scripts/src/commands/build.ts
@@ -42,7 +42,7 @@ export = async function({
     throw err;
   }
 
-  await applyHook(`before.${command}.load`, { args });
+  await applyHook(`before.${command}.load`, { args, webpackConfig: configArr });
 
   // eject config
   if (eject) {