+++ packages/build-scripts/test/pluginAPI.test.ts
@@ -1,5 +1,6 @@
 import Context from '../src/core/Context'
 import path = require('path')
+import Config = require('webpack-chain');
 
 describe('api regsiterMethod/applyMethod', () => {
   const context = new Context({
@@ -16,6 +17,8 @@ describe('api regsiterMethod/applyMethod', () => {
 
   it('api applyMethod', () => {
     const result = context.applyMethod('test', 'content')
+    expect(context.hasMethod('test')).toBe(true);
+    expect(context.hasMethod('unregistered')).toBe(false);
     expect(result).toBe('content')
   })
 
@@ -24,7 +27,31 @@ describe('api regsiterMethod/applyMethod', () => {
     expect(err instanceof Error).toBe(true)
   })
 
-})
+});
+
+describe('api regsiterTask/cancelTask', () => {
+  const context = new Context({
+    args: {},
+    command: 'start',
+    rootDir: path.join(__dirname, 'fixtures/basic/'),
+  });
+
+  it('api registerTask', async () => {
+    context.registerTask('task1', new Config());
+    const configArr = await context.setUp();
+    expect(configArr.length).toBe(1);
+  });
+
+  it('api cancelTask', async () => {
+    context.cancleTask('task1');
+    context.cancleTask('task2');
+    context.registerTask('task2', new Config());
+    context.registerTask('task3', new Config());
+    const configArr = await context.setUp();
+    expect(configArr.length).toBe(1);
+    expect(configArr[0].name).toBe('task3');
+  });
+});
 
 describe('api modifyUserConfig', () => {
   const context = new Context({